Projectile chanhes and Addition of levelviewlevelone
changes:
Introduce LevelViewLevelOne to serve as a dedicated UI class for Level One, inheriting from LevelView.This class does not add new functionality initially but acts as a placeholder for any future UI customizations specific to Level One.
Updated the LevelOne class to use LevelViewLevelOne instead of LevelView. This ensures that any Level One-specific UI changes can be easily integrated in the future.
2) edited enemy projectile and user projectile images to make thme better looking anf making killimg more effecient.